+++ desktop/e2e/example.spec.ts
@@ -1,23 +1,6 @@
-import { test, expect, _electron as electron } from "@playwright/test";
+import { expect } from "@playwright/test";
+import { test } from "./desktopE2EUtils";
 
-test("starts", async () => {
-  const app = await electron.launch({ args: [".vite/build/main.js"] });
-  const window = await app.firstWindow();
-  await window.waitForLoadState("domcontentloaded");
-  await expect(window.getByRole("heading", { name: "Bowser" })).toBeVisible();
-});
-
-test("can connect to server", async () => {
-  const app = await electron.launch({ args: [".vite/build/main.js"] });
-  const window = await app.firstWindow();
-  await window.waitForLoadState("domcontentloaded");
-
-  await window.getByLabel("Server address").fill("http://localhost:3000");
-  await window.getByLabel("Server Password").fill("aaa");
-
-  await window.getByRole("button", { name: "Connect" }).click();
-
-  await expect(
-    window.getByRole("heading", { name: "Select a show" }),
-  ).toBeVisible();
+test("smoke", async ({ app: [_app, win] }) => {
+  await expect(win.getByRole("heading", { name: "Bowser" })).toBeVisible();
 });
